resolved conflicts after merge master and activated publish project form
